BlockChain Dev/chatApplication/pyproject.toml
[tool.poetry]
name = "chatApplication"
version = "0.1.0"
description = ""
authors = ["sagarmohanty2k00 <sagarmohanty2k00@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
asyncio = "^3.4.3"
PyNaCl = "^1.5.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
